requirements.txt
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
Flask==3.1.1
gunicorn==21.2.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lyricsgenius==3.7.0
MarkupSafe==3.0.2
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.4
soupsieve==2.7
spotipy==2.25.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
